SSL/TLS Handshake speed up on Windows CE
When testing on Windows CE and SecureBlackBox V16 found the SSL handshake could take up to 75 seconds on low level x86 platform(800mhz).
This was due to the SecureBlackBox cypto provider being totally written in managed code.  When using a certificate from the Windows store it uses a different provider that uses the Windows Cypto API. In our case the handshake time went from 75 seconds to 0.4 sec.